DOS/api-vales/service/requirements.txt
psycopg2
grpcio
grpcio-tools
